#538 Add missing imports for a collection/map implementation and the parameter types
