+++ processor/src/test/java/org/mapstruct/ap/test/collection/map/MapMappingTest.java
@@ -18,9 +18,6 @@
  */
 package org.mapstruct.ap.test.collection.map;
 
-import static org.fest.assertions.Assertions.assertThat;
-import static org.fest.assertions.MapAssert.entry;
-
 import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.HashMap;
@@ -28,16 +25,21 @@ import java.util.Map;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mapstruct.ap.test.collection.map.other.ImportedType;
 import org.mapstruct.ap.testutil.IssueKey;
 import org.mapstruct.ap.testutil.WithClasses;
 import org.mapstruct.ap.testutil.runner.AnnotationProcessorTestRunner;
+
+import static org.fest.assertions.Assertions.assertThat;
+
+import static org.fest.assertions.MapAssert.entry;
 
 /**
  * Test for implementation of {@code Map} mapping methods.
  *
  * @author Gunnar Morling
  */
-@WithClasses({ SourceTargetMapper.class, CustomNumberMapper.class, Source.class, Target.class })
+@WithClasses({ SourceTargetMapper.class, CustomNumberMapper.class, Source.class, Target.class, ImportedType.class })
 @IssueKey("44")
 @RunWith(AnnotationProcessorTestRunner.class)
 public class MapMappingTest {
