+++ processor/src/test/java/org/mapstruct/ap/test/collection/map/Source.java
@@ -19,11 +19,15 @@
 package org.mapstruct.ap.test.collection.map;
 
 import java.util.Date;
+import java.util.LinkedHashMap;
 import java.util.Map;
+
+import org.mapstruct.ap.test.collection.map.other.ImportedType;
 
 public class Source {
 
     private Map<Long, Date> values;
+    private LinkedHashMap<String, ImportedType> stringEnumMap;
 
     public Map<Long, Date> getValues() {
         return values;
@@ -31,5 +35,13 @@ public class Source {
 
     public void setValues(Map<Long, Date> values) {
         this.values = values;
+    }
+
+    public LinkedHashMap<String, ImportedType> getStringEnumMap() {
+        return stringEnumMap;
+    }
+
+    public void setStringEnumMap(LinkedHashMap<String, ImportedType> stringEnumMap) {
+        this.stringEnumMap = stringEnumMap;
     }
 }
